refactor(ui): 拆分面板模块并稳定贴图弹窗
