--- views/converter/style-panel.js
@@ -1,1527 +1,0 @@
-/*
-## 核心功能
-
-实现转换器主面板的 style panel 交互能力。
-
-## 输入
-
-接收 AppleStyleView 实例状态、活动笔记、渲染结果、用户点击和面板控件事件。
-
-## 输出
-
-输出 `stylePanelMethods`，驱动预览刷新、样式选择、剪贴板或 AI layout 面板行为。
-
-## 定位
-
-位于 views/converter/，只处理转换器视图交互；底层转换和同步逻辑调用 services/。
-
-## 依赖
-
-关键依赖：`../apple-style-view-shared.js`。
-
-## 维护规则
-
-- 修改逻辑后同步更新本文件说明书，并检查 views/converter 的文件夹 README 是否仍准确。
-- 保持职责边界清晰，跨层行为优先通过既有服务、视图或测试 helper 协作。
-*/
-
-/* eslint-disable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-argument -- reason: Obsidian view state and DOM helpers are dynamically composed across method groups */
-
-import {
-  normalizeVaultPath,
-  createHtmlContainer,
-  toReadableError,
-  toRecord,
-  removeElementClass,
-  APPLE_STYLE_VIEW_TITLE,
-  getObsidianSetIcon,
-  getAppleThemeApi,
-  getEventTargetValue,
-  getImageSwipeCommandCopy,
-  isMobileClient,
-  resolveMarkdownSource,
-  MarkdownView,
-} from '../apple-style-view-shared.js';
-import {
-  extractStickerData,
-  STICKER_MAX_IMAGES,
-  STICKER_MAX_CONTENT_LENGTH,
-} from '../../services/sticker-extractor.js';
-import { getStickerImageItemSrc } from '../../services/sticker-image-items.js';
-import {
-  moveStickerImageItem,
-  renderStickerImageList,
-} from '../shared/sticker-image-list.js';
-
-/** @type {StylePanelMethodsContract & ThisType<AppleStyleViewContract>} */
-export const stylePanelMethods = {
-createSettingsPanel(container) {
-
-  // 1. 创建顶部双行 Header 容器
-  const header = container.createEl('div', { cls: 'apple-preview-header' });
-
-  // 1.1 第一行：Title Row
-  const topRow = header.createEl('div', { cls: 'apple-preview-header-top apple-top-toolbar' });
-  this.currentDocLabel = topRow.createEl('div', { cls: 'apple-toolbar-title' });
-  if (!isMobileClient(this.app)) {
-    const pluginLine = this.currentDocLabel.createDiv({ cls: 'apple-toolbar-plugin-line' });
-    pluginLine.createEl('span', { text: APPLE_STYLE_VIEW_TITLE, cls: 'apple-toolbar-plugin-name' });
-  }
-  this.docTitleText = this.currentDocLabel.createDiv({ text: '未选择文档', cls: 'apple-toolbar-doc-name' });
-
-  // 1.2 第二行：Action Row (初始隐藏，由 updateCurrentDoc 根据是否有选定文档展现)
-  const bottomRow = header.createEl('div', { cls: 'apple-preview-header-bottom hidden' });
-  this.headerBottomRow = bottomRow;
-
-  // 1.2.1 第二行左侧：微型 Icon 分段模式切换胶囊
-  const segment = bottomRow.createEl('div', { cls: 'apple-preview-mode-segment' });
-  const setIcon = getObsidianSetIcon();
-
-  const selfRec = toRecord(this);
-  
-  selfRec.btnArticleMode = segment.createEl('button', {
-    cls: 'apple-mode-btn active',
-    attr: { 'aria-label': '文章排版模式', 'title': '文章排版模式' }
-  });
-  const btnArt = /** @type {HTMLElement} */ (selfRec.btnArticleMode);
-  if (typeof setIcon === 'function') {
-    setIcon(btnArt, 'align-left');
-  }
-
-  selfRec.btnStickerMode = segment.createEl('button', {
-    cls: 'apple-mode-btn',
-    attr: { 'aria-label': '微信贴图模式', 'title': '微信贴图模式' }
-  });
-  const btnStk = /** @type {HTMLElement} */ (selfRec.btnStickerMode);
-  if (typeof setIcon === 'function') {
-    setIcon(btnStk, 'layout-grid');
-  }
-
-  btnArt.addEventListener('click', () => this.switchPreviewMode('article'));
-  btnStk.addEventListener('click', () => this.switchPreviewMode('sticker'));
-
-  // 1.2.2 第二行右侧：操作按钮组
-  const actions = bottomRow.createEl('div', { cls: 'apple-toolbar-actions' });
-
-  // 按钮工厂函数
-  /**
-   * @param {string} icon
-   * @param {string} title
-   * @param {() => unknown} onClick
-   * @returns {ObsidianElementLike}
-   */
-  const createIconBtn = (icon, title, onClick) => {
-    const btn = actions.createEl('div', {
-      cls: 'apple-icon-btn',
-      attr: { 'aria-label': title, 'title': title } // Tooltip
-    });
-    if (typeof setIcon === 'function') {
-      setIcon(btn, icon);
-    }
-    btn.addEventListener('click', onClick);
-    return btn;
-  };
-
-  // [设置] 按钮
-  const settingsButton = createIconBtn('sliders-horizontal', '公众号排版样式设置', () => {
-    this.toggleSettingsPanel();
-  });
-  this.settingsBtn = settingsButton;
-
-  this.aiLayoutBtn = createIconBtn('sparkles', 'AI 编排', () => this.onAiLayoutButtonClick());
-
-  // [复制] 按钮（移动端隐藏，避免误导）
-  if (!isMobileClient(this.app)) {
-    this.copyBtn = createIconBtn('copy', '复制到公众号', () => this.copyHTML());
-  } else {
-    this.copyBtn = null;
-  }
-
-  // [同步] 按钮（始终显示；未配置账号时点击后引导去设置）
-  createIconBtn('send', '发布与分发', () => this.showSyncModal());
-
-  // 2. 创建悬浮设置层 (初始隐藏)
-  this.settingsOverlay = container.createEl('div', { cls: 'apple-settings-overlay' });
-  const settingsArea = this.settingsOverlay.createEl('div', { cls: 'apple-settings-area' });
-  this.settingsArea = settingsArea;
-
-  const articleSettingsWrapper = /** @type {ObsidianElementLike} */ (settingsArea.createEl('div', { cls: 'apple-settings-article-wrapper' }));
-  this.articleSettingsWrapper = articleSettingsWrapper;
-
-  const stickerSettingsWrapper = /** @type {ObsidianElementLike} */ (settingsArea.createEl('div', { cls: 'apple-settings-sticker-wrapper hidden' }));
-  this.stickerSettingsWrapper = stickerSettingsWrapper;
-
-  const targetArea = articleSettingsWrapper;
-
-  // === 主题选择 ===
-  this.createSection(targetArea, '主题', (section) => {
-    const grid = section.createEl('div', { cls: 'apple-btn-grid' });
-    const themes = getAppleThemeApi().getThemeList();
-    themes.forEach(t => {
-      const btn = grid.createEl('button', {
-        cls: `apple-btn-theme ${this.plugin.settings.theme === t.value ? 'active' : ''}`,
-        text: t.label,
-        attr: { title: t.label },
-      });
-      btn.dataset.value = t.value;
-      btn.addEventListener('click', () => this.onThemeChange(t.value, grid));
-    });
-  });
-
-  // === 字体选择 ===
-  this.createSection(targetArea, '字体', (section) => {
-    const select = /** @type {ObsidianInputLike} */ (section.createEl('select', { cls: 'apple-select' }));
-    [
-      { value: 'sans-serif', label: '无衬线' },
-      { value: 'serif', label: '衬线' },
-      { value: 'monospace', label: '等宽' },
-    ].forEach(opt => {
-      const option = /** @type {ObsidianInputLike} */ (select.createEl('option', { value: opt.value, text: opt.label }));
-      if (this.plugin.settings.fontFamily === opt.value) option.selected = true;
-    });
-    select.addEventListener('change', (e) => this.onFontFamilyChange(getEventTargetValue(e, this.plugin.settings.fontFamily)));
-  });
-
-  // === 字号选择 ===
-  this.createSection(targetArea, '字号', (section) => {
-    const grid = section.createEl('div', { cls: 'apple-btn-row' });
-    const sizeOpts = [
-      { value: 1, label: '小' },
-      { value: 2, label: '较小' },
-      { value: 3, label: '推荐' },
-      { value: 4, label: '较大' },
-      { value: 5, label: '大' },
-    ];
-
-    sizeOpts.forEach(s => {
-      const btn = grid.createEl('button', {
-        cls: `apple-btn-size ${this.plugin.settings.fontSize === s.value ? 'active' : ''}`,
-        text: s.label,
-      });
-      btn.dataset.value = s.value;
-      btn.addEventListener('click', () => this.onFontSizeChange(s.value, grid));
-    });
-  });
-
-  // === 主题色 (移到标题样式上方) ===
-  this.createSection(targetArea, '主题色', (section) => {
-    const grid = section.createEl('div', { cls: 'apple-color-grid' });
-    const colors = getAppleThemeApi().getColorList();
-
-    // 预设颜色
-    colors.forEach(c => {
-      const btn = grid.createEl('button', {
-        cls: `apple-btn-color ${this.plugin.settings.themeColor === c.value ? 'active' : ''}`,
-      });
-      btn.dataset.value = c.value;
-      btn.style.setProperty('--btn-color', c.color);
-      btn.addEventListener('click', () => this.onColorChange(c.value, grid));
-    });
-
-    // 自定义颜色
-    const customBtn = grid.createEl('button', {
-      cls: `apple-btn-custom-text ${this.plugin.settings.themeColor === 'custom' ? 'active' : ''}`,
-      text: '自定义',
-      title: '自定义颜色'
-    });
-    customBtn.dataset.value = 'custom';
-
-    // 隐藏的颜色选择器
-    const colorInput = /** @type {ObsidianInputLike} */ (grid.createEl('input', {
-      type: 'color',
-      cls: 'apple-color-picker-hidden'
-    }));
-    colorInput.value = this.plugin.settings.customColor || '#000000';
-    colorInput.setCssStyles({
-      visibility: 'hidden',
-      width: '0',
-      height: '0',
-      position: 'absolute',
-    });
-
-    // 点击按钮触发颜色选择
-    customBtn.addEventListener('click', () => {
-      colorInput.click();
-    });
-
-    // 颜色改变实时预览
-    colorInput.addEventListener('input', (e) => {
-      customBtn.style.setProperty('--btn-color', getEventTargetValue(e, this.plugin.settings.customColor));
-    });
-
-    // 颜色确认后保存
-    colorInput.addEventListener('change', async (e) => {
-      const newColor = getEventTargetValue(e, this.plugin.settings.customColor);
-      customBtn.style.setProperty('--btn-color', newColor);
-
-      // 更新设置
-      this.plugin.settings.customColor = newColor;
-      this.theme.update({ customColor: newColor });
-      await this.onColorChange('custom', grid);
-    });
-  });
-
-  // === 页面两侧留白 ===
-  this.createSection(targetArea, '页面两侧留白', (section) => {
-    const mobile = isMobileClient(this.app);
-    const container = section.createEl('div', {
-      cls: 'apple-slider-container',
-      style: 'width: 100%; display: flex; align-items: center; gap: 10px;'
-    });
-
-    const slider = /** @type {ObsidianInputLike} */ (container.createEl('input', {
-      type: 'range',
-      cls: 'apple-slider',
-      attr: { min: 0, max: mobile ? 36 : 40, step: 1 }
-    }));
-    slider.value = this.plugin.settings.sidePadding;
-    slider.setCssStyles({ flex: '1' });
-
-    const valueLabel = container.createEl('span', {
-      text: `${this.plugin.settings.sidePadding}px`,
-      style: 'font-size: 12px; color: var(--apple-secondary); min-width: 32px; text-align: right;'
-    });
-
-    slider.addEventListener('input', (e) => {
-      const val = parseInt(getEventTargetValue(e, String(this.plugin.settings.sidePadding)), 10);
-      valueLabel.setText(`${val}px`);
-      // 拖动过程中只做轻量更新，避免移动端手势被重渲染卡住。
-      this.plugin.settings.sidePadding = val;
-      this.theme.update({ sidePadding: val });
-
-      if (this.saveTimeout) window.clearTimeout(this.saveTimeout);
-      this.saveTimeout = window.setTimeout(async () => {
-        await this.plugin.saveSettings();
-      }, 500);
-      this.scheduleSidePaddingPreview(mobile ? 220 : 120);
-    });
-
-    slider.addEventListener('change', async (e) => {
-      const val = parseInt(getEventTargetValue(e, String(this.plugin.settings.sidePadding)), 10);
-      valueLabel.setText(`${val}px`);
-      this.plugin.settings.sidePadding = val;
-      this.theme.update({ sidePadding: val });
-      if (this.sidePaddingPreviewTimer) {
-        window.clearTimeout(this.sidePaddingPreviewTimer);
-        this.sidePaddingPreviewTimer = null;
-      }
-      await this.plugin.saveSettings();
-      await this.convertCurrent(true);
-    });
-  });
-
-  // === 排版间距（折叠：呼吸感微调，默认收起） ===
-  const spacingGroup = targetArea.createEl('details', { cls: 'apple-settings-details' });
-  this.settingsSpacingGroup = spacingGroup;
-  const spacingSummary = spacingGroup.createEl('summary', { cls: 'apple-settings-summary' });
-  spacingSummary.createEl('span', { text: '排版间距' });
-  this.settingsSpacingValues = spacingSummary.createEl('span', {
-    attr: {
-      style: 'margin-left: auto; margin-right: 8px; font-size: 11px; font-weight: 400; color: var(--apple-secondary);'
-    }
-  });
-  const spacingArea = spacingGroup.createDiv({ cls: 'apple-settings-area apple-settings-advanced-area' });
-  this.updateSpacingSummary();
-
-  /**
-   * 复用留白滑块交互：input 轻量预览 + 防抖保存；change 全量重渲染 + 保存。
-   * @param {string} label
-   * @param {number} min
-   * @param {number} max
-   * @param {number} step
-   * @param {() => number} getEffective
-   * @param {string} settingsKey
-   * @param {string} updateKey
-   */
-  const buildSpacingSlider = (label, min, max, step, getEffective, settingsKey, updateKey) => {
-    this.createSection(spacingArea, label, (section) => {
-      const container = section.createEl('div', {
-        cls: 'apple-slider-container',
-        style: 'width: 100%; display: flex; align-items: center; gap: 10px;'
-      });
-
-      const slider = /** @type {ObsidianInputLike} */ (container.createEl('input', {
-        type: 'range',
-        cls: 'apple-slider',
-        attr: { min: String(min), max: String(max), step: String(step) }
-      }));
-      const initial = getEffective();
-      slider.value = String(initial);
-      slider.setCssStyles({ flex: '1' });
-
-      const valueLabel = container.createEl('span', {
-        text: this.formatSpacingValue(initial),
-        style: 'font-size: 12px; color: var(--apple-secondary); min-width: 32px; text-align: right;'
-      });
-
-      // 记录引用，供 refreshSpacingSliders() 切主题后刷新显示。
-      this.spacingSliderRefs.push({ slider, valueLabel, getEffective });
-
-      const applyValue = (raw) => {
-        const val = Number(raw);
-        valueLabel.setText(this.formatSpacingValue(val));
-        this.plugin.settings[settingsKey] = val;
-        this.theme.update({ [updateKey]: val });
-        this.updateSpacingSummary();
-      };
-
-      slider.addEventListener('input', (e) => {
-        const val = Number(getEventTargetValue(e, String(getEffective())));
-        applyValue(val);
-        if (this.saveTimeout) window.clearTimeout(this.saveTimeout);
-        this.saveTimeout = window.setTimeout(async () => {
-          await this.plugin.saveSettings();
-        }, 500);
-        this.scheduleSidePaddingPreview(120);
-      });
-
-      slider.addEventListener('change', async (e) => {
-        const val = Number(getEventTargetValue(e, String(getEffective())));
-        applyValue(val);
-        if (this.sidePaddingPreviewTimer) {
-          window.clearTimeout(this.sidePaddingPreviewTimer);
-          this.sidePaddingPreviewTimer = null;
-        }
-        await this.plugin.saveSettings();
-        await this.convertCurrent(true);
-      });
-    });
-  };
-
-  // 存储滑块引用，供切主题时刷新显示值（跟随主题的滑块需同步新主题默认值）。
-  this.spacingSliderRefs = [];
-  buildSpacingSlider('行间距', 1.4, 2.2, 0.05, () => this.getEffectiveLineHeight(), 'lineHeight', 'lineHeight');
-  buildSpacingSlider('段间距', 8, 40, 1, () => this.getEffectiveParagraphGap(), 'paragraphGap', 'paragraphGap');
-  buildSpacingSlider('字间距', 0, 2, 0.5, () => this.getEffectiveLetterSpacing(), 'letterSpacing', 'letterSpacing');
-
-  const advancedOptions = targetArea.createEl('details', { cls: 'apple-settings-details' });
-  this.settingsAdvancedOptions = advancedOptions;
-  advancedOptions.createEl('summary', {
-    cls: 'apple-settings-summary',
-    text: '高级选项'
-  });
-  const advancedArea = advancedOptions.createDiv({ cls: 'apple-settings-area apple-settings-advanced-area' });
-  this.settingsAdvancedArea = advancedArea;
-
-  // === 引用样式 ===
-  const quoteStyleSection = this.createSection(advancedArea, '引用样式', (section) => {
-    const select = /** @type {ObsidianInputLike} */ (section.createEl('select', { cls: 'apple-select' }));
-    [
-      { value: 'theme', label: '经典主题色' },
-      { value: 'neutral', label: '中性灰（推荐）' },
-    ].forEach((opt) => {
-      const option = /** @type {ObsidianInputLike} */ (select.createEl('option', { value: opt.value, text: opt.label }));
-      if (this.plugin.settings.quoteCalloutStyleMode === opt.value) option.selected = true;
-    });
-    select.addEventListener('change', (e) => this.onQuoteCalloutStyleModeChange(getEventTargetValue(e, this.plugin.settings.quoteCalloutStyleMode)));
-
-    section.createEl('span', {
-      text: '中性灰更适合长文阅读；经典主题色兼容现有风格。',
-      attr: {
-        style: 'font-size: 11px; color: var(--apple-secondary); margin-top: 8px; opacity: 0.8; font-weight: 500; display: block;'
-      }
-    });
-  });
-  quoteStyleSection.classList.add('apple-settings-featured');
-
-  // === 标题样式 (移到主题色下方) ===
-  const headingStyleSection = this.createSection(advancedArea, '标题样式', (section) => {
-    const row = section.createEl('div', { cls: 'apple-settings-inline-row' });
-
-    const toggle = row.createEl('label', { cls: 'apple-toggle' });
-    const checkbox = /** @type {ObsidianInputLike} */ (toggle.createEl('input', { type: 'checkbox', cls: 'apple-toggle-input' }));
-    checkbox.checked = this.plugin.settings.coloredHeader;
-    toggle.createEl('span', { cls: 'apple-toggle-slider' });
-
-    section.createEl('span', {
-      text: '标题使用加深主题色',
-      attr: {
-        style: 'font-size: 11px; color: var(--apple-secondary); opacity: 0.8; font-weight: 500; display: block;'
-      }
-    });
-
-    checkbox.addEventListener('change', async () => {
-      this.plugin.settings.coloredHeader = checkbox.checked;
-      await this.plugin.saveSettings();
-
-      // 关键修复：更新主题状态并重绘
-      this.theme.update({ coloredHeader: checkbox.checked });
-      // 强制刷新
-      await this.convertCurrent(true);
-    });
-  });
-  headingStyleSection.classList.add('apple-settings-inline-toggle');
-
-  // === 正文标点标准化 ===
-  const punctuationSection = this.createSection(advancedArea, '正文标点标准化', (section) => {
-    const row = section.createEl('div', { cls: 'apple-settings-inline-row' });
-    const toggle = row.createEl('label', { cls: 'apple-toggle' });
-    const checkbox = /** @type {ObsidianInputLike} */ (toggle.createEl('input', { type: 'checkbox', cls: 'apple-toggle-input' }));
-    checkbox.checked = this.plugin.settings.normalizeChinesePunctuation === true;
-    toggle.createEl('span', { cls: 'apple-toggle-slider' });
-
-    section.createEl('span', {
-      text: '仅作用于预览 / 复制 / 同步结果',
-      attr: {
-        style: 'font-size: 11px; color: var(--apple-secondary); opacity: 0.8; font-weight: 500; display: block;'
-      }
-    });
-
-    checkbox.addEventListener('change', async () => {
-      this.plugin.settings.normalizeChinesePunctuation = checkbox.checked;
-      await this.plugin.saveSettings();
-      await this.convertCurrent(true);
-    });
-  });
-  punctuationSection.classList.add('apple-settings-inline-toggle');
-
-  // === Mac 代码块开关 ===
-  const macCodeSection = this.createSection(advancedArea, 'Mac 风格代码块', (section) => {
-    const row = section.createEl('div', { cls: 'apple-settings-inline-row' });
-    const toggle = row.createEl('label', { cls: 'apple-toggle' });
-    const checkbox = /** @type {ObsidianInputLike} */ (toggle.createEl('input', { type: 'checkbox', cls: 'apple-toggle-input' }));
-    checkbox.checked = this.plugin.settings.macCodeBlock;
-    toggle.createEl('span', { cls: 'apple-toggle-slider' });
-    checkbox.addEventListener('change', () => this.onMacCodeBlockChange(checkbox.checked));
-  });
-  macCodeSection.classList.add('apple-settings-inline-toggle');
-
-  // === 代码块行号开关 ===
-  const codeLineNumberSection = this.createSection(advancedArea, '显示代码行号', (section) => {
-    const row = section.createEl('div', { cls: 'apple-settings-inline-row' });
-    const toggle = row.createEl('label', { cls: 'apple-toggle' });
-    const checkbox = /** @type {ObsidianInputLike} */ (toggle.createEl('input', { type: 'checkbox', cls: 'apple-toggle-input' }));
-    checkbox.checked = this.plugin.settings.codeLineNumber;
-    toggle.createEl('span', { cls: 'apple-toggle-slider' });
-    checkbox.addEventListener('change', () => this.onCodeLineNumberChange(checkbox.checked));
-  });
-  codeLineNumberSection.classList.add('apple-settings-inline-toggle');
-
-  // === 显示图片说明文字 ===
-  const captionSection = this.createSection(advancedArea, '显示图片说明文字', (section) => {
-    const row = section.createEl('div', { cls: 'apple-settings-inline-row' });
-    const toggle = row.createEl('label', { cls: 'apple-toggle' });
-    const checkbox = /** @type {ObsidianInputLike} */ (toggle.createEl('input', { type: 'checkbox', cls: 'apple-toggle-input' }));
-    checkbox.checked = this.plugin.settings.showImageCaption;
-    toggle.createEl('span', { cls: 'apple-toggle-slider' });
-
-    section.createEl('span', {
-      text: '关闭水印时，在图片下方显示说明文字',
-      attr: {
-        style: 'font-size: 11px; color: var(--apple-secondary); opacity: 0.8; font-weight: 500; display: block;'
-      }
-    });
-
-    checkbox.addEventListener('change', async () => {
-      this.plugin.settings.showImageCaption = checkbox.checked;
-      await this.plugin.saveSettings();
-
-      if (this.converter) {
-        this.converter.updateConfig({ showImageCaption: checkbox.checked });
-        await this.convertCurrent(true);
-      }
-    });
-
-    this.captionToggleState = { checkbox, toggle };
-  });
-  captionSection.classList.add('apple-settings-inline-toggle');
-
-  // === 横滑图片块提示 ===
-  this.createSection(advancedArea, '横滑图片块', (section) => {
-    const imageBlockCommand = getImageSwipeCommandCopy(this.app, 'image-swipe').name;
-    const sensitiveImageBlockCommand = getImageSwipeCommandCopy(this.app, 'image-sensitive').name;
-    section.createEl('span', {
-      text: `选中多张图片，打开命令面板，运行「${imageBlockCommand}」或「${sensitiveImageBlockCommand}」。`,
-      attr: {
-        style: 'font-size: 11px; color: var(--apple-secondary); opacity: 0.78; font-weight: 500; line-height: 1.6; display: block;'
-      }
-    });
-  });
-
-  // 根据全局水印设置更新状态
-  if (this.plugin.settings.enableWatermark) {
-    const captionDesc = captionSection.querySelector('.apple-setting-content > span');
-    if (captionDesc) {
-      captionDesc.setText('因全局设置中已开启水印，此选项默认开启');
-    }
-    const toggleState = this.captionToggleState;
-    if (toggleState?.checkbox) {
-      toggleState.checkbox.checked = true;
-      toggleState.checkbox.disabled = true;
-    }
-    if (toggleState?.toggle) {
-      toggleState.toggle.setCssStyles({
-        pointerEvents: 'none',
-        opacity: '0.6',
-        filter: 'grayscale(100%)',
-      });
-    }
-  }
-
-  // === 3. 渲染贴图模式专属设置 ===
-  const stickerWrapperRaw = /** @type {unknown} */ (this.stickerSettingsWrapper);
-  if (stickerWrapperRaw && typeof stickerWrapperRaw === 'object' && 'createDiv' in stickerWrapperRaw) {
-    const stickerWrapper = /** @type {ObsidianElementLike} */ (stickerWrapperRaw);
-    const stickerHeader = /** @type {ObsidianElementLike} */ (stickerWrapper.createDiv({ cls: 'apple-settings-sticker-header' }));
-    stickerHeader.createEl('span', { text: '微信贴图设置', cls: 'title' });
-
-    // 3.2 配图序号标注开关
-    const indexSection = this.createSection(stickerWrapper, '配图序号标注', (section) => {
-      const container = section.createEl('div', { cls: 'apple-sticker-toggle-row' });
-      container.createEl('span', { text: '正文插入配图序号', cls: 'apple-sticker-toggle-label' });
-
-      const toggle = container.createDiv({ cls: 'apple-toggle' });
-      const checkbox = toggle.createEl('input', { type: 'checkbox', cls: 'apple-toggle-input' });
-      checkbox.checked = Boolean(this.insertStickerImageIndex);
-      toggle.createEl('span', { cls: 'apple-toggle-slider' });
-
-      container.addEventListener('click', (e) => {
-        if (e.target !== checkbox) {
-          e.preventDefault();
-          checkbox.checked = !checkbox.checked;
-          checkbox.dispatchEvent(new Event('change'));
-        }
-      });
-
-      checkbox.addEventListener('change', () => {
-        this.insertStickerImageIndex = checkbox.checked;
-        this.renderStickerPreview();
-      });
-      this.stickerIndexToggleState = { checkbox, toggle };
-    });
-    indexSection.classList.add('apple-settings-inline-toggle');
-
-    // 3.3 过滤说明
-    this.createSection(stickerWrapper, '排版优化说明', (section) => {
-      const card = section.createDiv({ cls: 'apple-settings-info-card' });
-      card.createEl('p', {
-        text: '· 自动忽略 Frontmatter YAML 头信息\n· 自动剥离不支持的代码块与表格以保证文案纯净\n· 支持拖拽更改贴图图片顺序',
-        cls: 'apple-settings-info-card-text'
-      });
-    });
-  }
-
-  this.aiLayoutOverlay = container.createEl('div', { cls: 'apple-ai-layout-overlay' });
-  this.createAiLayoutPanel(this.aiLayoutOverlay);
-  this.updateAiToolbarState();
-}
-,
-
-createAccountSelector(parent) {
-  /** @type {WechatAccountLike[]} */
-  const accounts = this.plugin.settings.wechatAccounts || [];
-  if (accounts.length === 0) return;
-
-  const section = parent.createEl('div', { cls: 'apple-setting-section wechat-account-selector' });
-  section.createEl('label', { cls: 'apple-setting-label', text: '同步账号' });
-
-  const select = /** @type {ObsidianInputLike} */ (section.createEl('select', { cls: 'wechat-account-select' }));
-
-  const defaultId = this.plugin.settings.defaultAccountId;
-
-  for (const account of accounts) {
-    const option = /** @type {ObsidianInputLike} */ (select.createEl('option', {
-      value: account.id,
-      text: account.id === defaultId ? `${account.name} (默认)` : account.name
-    }));
-    if (account.id === defaultId) {
-      option.selected = true;
-    }
-  }
-
-  // 保存选中的账号 ID 到实例属性
-  this.selectedAccountId = defaultId;
-  select.addEventListener('change', (event) => {
-    this.selectedAccountId = getEventTargetValue(event, defaultId);
-  });
-}
-,
-
-getFirstImageFromArticle() {
-  if (!this.currentHtml) return null;
-  const tempDiv = createHtmlContainer('div', this.currentHtml);
-  const imgs = Array.from(tempDiv.querySelectorAll('img'));
-
-  // 遍历所有图片，跳过头像（alt="logo"）
-  for (const img of imgs) {
-    if (img.alt === 'logo') continue;
-    const src = String(img.getAttribute('src') || img.src || '').trim();
-    if (src) return src;
-  }
-  return null;
-}
-,
-
-getPublishContextFile() {
-  const activeFile = this.app?.workspace?.getActiveFile?.();
-  if (activeFile) return activeFile;
-  if (this.lastActiveFile) return this.lastActiveFile;
-  return null;
-}
-,
-
-getFrontmatterPublishMeta(activeFile) {
-  if (!activeFile) {
-    return { excerpt: '', cover: '', cover_dir: '', coverSrc: null, title: '' };
-  }
-
-  const frontmatter = this.app?.metadataCache?.getFileCache?.(activeFile)?.frontmatter;
-  const excerpt = this.getFrontmatterString(frontmatter, ['excerpt']);
-  const cover = this.getFrontmatterString(frontmatter, ['cover']);
-  const cover_dir = this.getFrontmatterString(frontmatter, ['cover_dir', 'coverDir', 'cover-dir', 'coverdir', 'CoverDIR']);
-  const title = this.getFrontmatterString(frontmatter, ['title']);
-
-  // 解析失败时静默回退：返回 null，不中断流程
-  const coverSrc = cover ? this.resolveVaultPathToResourceSrc(cover) : null;
-
-  return { excerpt, cover, cover_dir, coverSrc, title };
-}
-,
-
-getFrontmatterString(frontmatter, keys) {
-  const frontmatterRecord = toRecord(frontmatter);
-  if (!frontmatterRecord) return '';
-  if (!Array.isArray(keys) || keys.length === 0) return '';
-
-  const normalizedTargets = new Set(keys.map(key => this.normalizeFrontmatterKey(key)));
-  for (const key of keys) {
-    const value = frontmatterRecord[key];
-    if (typeof value === 'string' && value.trim()) return value.trim();
-  }
-
-  for (const [key, value] of Object.entries(frontmatterRecord)) {
-    if (!normalizedTargets.has(this.normalizeFrontmatterKey(key))) continue;
-    if (typeof value === 'string' && value.trim()) return value.trim();
-  }
-
-  return '';
-}
-,
-
-normalizeFrontmatterKey(key) {
-  return String(key || '').toLowerCase().replace(/[_-]/g, '');
-}
-,
-
-getFrontmatterKeyMap(frontmatter, keys) {
-  /** @type {Record<string, string>} */
-  const result = {};
-  const frontmatterRecord = toRecord(frontmatter);
-  if (!frontmatterRecord) return result;
-  if (!Array.isArray(keys) || keys.length === 0) return result;
-
-  const normalizedTargets = new Set(keys.map(key => this.normalizeFrontmatterKey(key)));
-  for (const [key, value] of Object.entries(frontmatterRecord)) {
-    if (!normalizedTargets.has(this.normalizeFrontmatterKey(key))) continue;
-    if (typeof value !== 'string') continue;
-    const normalizedValue = this.normalizeVaultPath(value);
-    if (!normalizedValue) continue;
-    result[key] = normalizedValue;
-  }
-  return result;
-}
-,
-
-/** @param {string} filePath @param {string} dirPath @returns {boolean} */
-isPathInsideDirectory(filePath, dirPath) {
-  const file = this.normalizeVaultPath(filePath);
-  const dir = this.normalizeVaultPath(dirPath);
-  if (!file || !dir) return false;
-  if (file === dir) return true;
-  return file.startsWith(`${dir}/`);
-}
-,
-
-/** @param {string} filePath @param {string} dirPath @returns {boolean} */
-isPathInsideDirectoryByTail(filePath, dirPath) {
-  const file = this.normalizeVaultPath(filePath);
-  const dir = this.normalizeVaultPath(dirPath);
-  if (!file || !dir) return false;
-
-  const dirSegments = dir.split('/').filter(Boolean);
-  if (dirSegments.length < 2) return false;
-
-  // 允许清理目录与 frontmatter 路径存在“根前缀差异”
-  // 例如 cleanedDir: Wechat/published/img
-  //      cover:     published/img/post-cover.jpg
-  for (let i = 1; i <= dirSegments.length - 2; i++) {
-    const tailDir = dirSegments.slice(i).join('/');
-    if (this.isPathInsideDirectory(file, tailDir)) {
-      return true;
-    }
-  }
-  return false;
-}
-,
-
-/** @param {string} pathValue @param {string} cleanedDir @returns {boolean} */
-shouldClearFrontmatterPathAfterCleanup(pathValue, cleanedDir) {
-  const normalized = this.normalizeVaultPath(pathValue);
-  if (!normalized) return false;
-  if (this.isPathInsideDirectory(normalized, cleanedDir)) return true;
-  return this.isPathInsideDirectoryByTail(normalized, cleanedDir);
-}
-,
-
-clearInvalidPublishMetaInFrontmatter(frontmatter, cleanedDir) {
-  const frontmatterRecord = toRecord(frontmatter);
-  if (!frontmatterRecord) return false;
-
-  let changed = false;
-  const coverMap = this.getFrontmatterKeyMap(frontmatter, ['cover']);
-  const coverDirMap = this.getFrontmatterKeyMap(frontmatter, ['cover_dir', 'coverDir', 'cover-dir', 'coverdir', 'CoverDIR']);
-
-  for (const [key, value] of Object.entries(coverMap)) {
-    if (this.shouldClearFrontmatterPathAfterCleanup(value, cleanedDir)) {
-      frontmatterRecord[key] = '';
-      changed = true;
-    }
-  }
-
-  for (const [key, value] of Object.entries(coverDirMap)) {
-    if (this.shouldClearFrontmatterPathAfterCleanup(value, cleanedDir)) {
-      frontmatterRecord[key] = '';
-      changed = true;
-    }
-  }
-
-  return changed;
-}
-,
-
-/** @param {TFileLike} activeFile @param {string} cleanedDir @returns {Promise<boolean>} */
-async clearInvalidPublishMetaByTextFallback(activeFile, cleanedDir) {
-  const vault = this.app?.vault;
-  if (!vault || typeof vault.read !== 'function' || typeof vault.modify !== 'function') {
-    return false;
-  }
-
-  const source = await vault.read(activeFile);
-  if (typeof source !== 'string' || !source.startsWith('---')) return false;
-
-  const match = source.match(/^(---[ \t]*\r?\n)([\s\S]*?)(\r?\n(?:---|\.\.\.)[ \t]*(?:\r?\n|$))/);
-  if (!match) return false;
-
-  let changed = false;
-  const body = match[2].replace(/^([ \t]*)(cover|cover_dir|coverDir|cover-dir|coverdir|CoverDIR)([ \t]*:[ \t]*)(.*)$/gmi, (line, indent, key, separator, rawValue) => {
-    const value = String(rawValue || '').trim().replace(/^['"]|['"]$/g, '');
-    if (!this.shouldClearFrontmatterPathAfterCleanup(value, cleanedDir)) {
-      return line;
-    }
-    changed = true;
-    return `${indent}${key}${separator}''`;
-  });
-
-  if (!changed) return false;
-  await vault.modify(activeFile, `${match[1]}${body}${match[3]}${source.slice(match[0].length)}`);
-  return true;
-}
-,
-
-/** @param {TFileLike | null | undefined} activeFile @param {string} cleanedDirPath @returns {Promise<string | null>} */
-async clearInvalidPublishMetaAfterCleanup(activeFile, cleanedDirPath) {
-  if (!activeFile || !cleanedDirPath) return null;
-
-  const cleanedDir = this.normalizeVaultPath(cleanedDirPath);
-  if (!cleanedDir) return null;
-
-  try {
-    const processFrontMatter = this.app?.fileManager?.['processFrontMatter'];
-    if (typeof processFrontMatter === 'function') {
-      await processFrontMatter.call(this.app.fileManager, activeFile, (frontmatter) => {
-        this.clearInvalidPublishMetaInFrontmatter(toRecord(frontmatter), cleanedDir);
-      });
-    } else {
-      await this.clearInvalidPublishMetaByTextFallback(activeFile, cleanedDir);
-    }
-  } catch (error) {
-    return `资源已删除，但清理 frontmatter 中失效的 cover/cover_dir 失败: ${toReadableError(error).message}`;
-  }
-
-  return null;
-}
-,
-
-/** @param {unknown} vaultPath @returns {string | null} */
-resolveVaultPathToResourceSrc(vaultPath) {
-  if (typeof vaultPath !== 'string') return null;
-  const normalized = vaultPath.trim().replace(/\\/g, '/').replace(/^\/+/, '');
-  if (!normalized) return null;
-
-  try {
-    const file = this.app.vault.getAbstractFileByPath(normalized);
-    if (!file) return null;
-    if (typeof file.extension !== 'string') return null; // 仅接受文件，不接受目录
-    return this.app.vault.getResourcePath(file);
-  } catch {
-    // frontmatter 路径失效或不是文件时，静默回退
-    return null;
-  }
-}
-,
-
-/** @param {unknown} vaultPath @returns {string} */
-normalizeVaultPath(vaultPath) {
-  return normalizeVaultPath(vaultPath);
-}
-,
-
-/** @returns {string} */
-getVaultConfigDir() {
-  const configDir = this.app?.vault?.configDir;
-  return typeof configDir === 'string' ? this.normalizeVaultPath(configDir) : '';
-}
-,
-
-/** @returns {string} */
-getCleanupDirTemplate() {
-  const raw = typeof this.plugin?.settings?.cleanupDirTemplate === 'string'
-    ? this.plugin.settings.cleanupDirTemplate
-    : '';
-  return this.normalizeVaultPath(raw);
-}
-,
-
-/** @param {TFileLike | null | undefined} activeFile @returns {{ path: string, warning?: string }} */
-resolveCleanupDirPath(activeFile) {
-  const template = this.getCleanupDirTemplate();
-  if (!template) {
-    return { path: '', warning: '未配置清理目录，请在插件设置中先填写目录后再启用自动清理' };
-  }
-
-  const hasNotePlaceholder = /\{\{\s*note\s*\}\}/i.test(template);
-  if (hasNotePlaceholder && !activeFile) {
-    return { path: '', warning: '当前没有活动文档，无法解析清理目录中的 {{note}}' };
-  }
-
-  const noteName = typeof activeFile?.basename === 'string' ? activeFile.basename.trim() : '';
-  const resolved = template.replace(/\{\{\s*note\s*\}\}/gi, noteName);
-  const normalized = this.normalizeVaultPath(resolved);
-  if (!normalized) {
-    return { path: '', warning: '清理目录为空，请检查设置值' };
-  }
-
-  return { path: normalized };
-}
-,
-
-/** @param {string} vaultPath @returns {boolean} */
-isSafeCleanupDirPath(vaultPath) {
-  const normalized = this.normalizeVaultPath(vaultPath);
-  if (!normalized) return false;
-  if (normalized === '.') return false;
-  if (normalized.includes('..')) return false;
-  const configDir = this.getVaultConfigDir();
-  if (configDir && (normalized === configDir || normalized.startsWith(`${configDir}/`))) return false;
-  return true;
-}
-,
-
-async cleanupConfiguredDirectory(activeFile) {
-  if (!this.plugin.settings.cleanupAfterSync) {
-    return { attempted: false };
-  }
-
-  const useSystemTrash = this.plugin.settings.cleanupUseSystemTrash !== false;
-  const resolved = this.resolveCleanupDirPath(activeFile);
-  if (!resolved.path) {
-    return { attempted: true, success: false, warning: resolved.warning || '未解析到清理目录' };
-  }
-
-  const normalized = resolved.path;
-  if (!this.isSafeCleanupDirPath(normalized)) {
-    return { attempted: true, success: false, warning: `清理目录不安全，已跳过: ${normalized}` };
-  }
-
-  const abstractFile = this.app.vault.getAbstractFileByPath(normalized);
-  if (!abstractFile) {
-    return { attempted: true, success: false, warning: `清理目录不存在: ${normalized}` };
-  }
-
-  const isFile = typeof abstractFile.extension === 'string';
-  if (isFile) {
-    return { attempted: true, success: false, warning: `清理路径不是目录，已跳过: ${normalized}` };
-  }
-
-  try {
-    if (typeof this.app.vault.trash === 'function') {
-      await this.app.vault.trash(abstractFile, useSystemTrash);
-    } else if (typeof this.app.vault.delete === 'function') {
-      await this.app.vault.delete(abstractFile, true);
-    } else {
-      throw new Error('当前 Obsidian 版本不支持删除接口');
-    }
-  } catch (error) {
-    return { attempted: true, success: false, warning: `删除失败 (${normalized}): ${toReadableError(error).message}` };
-  }
-
-  const frontmatterWarning = await this.clearInvalidPublishMetaAfterCleanup(activeFile, normalized);
-  if (frontmatterWarning) {
-    return { attempted: true, success: true, cleanedPath: normalized, warning: frontmatterWarning };
-  }
-
-  return { attempted: true, success: true, cleanedPath: normalized };
-}
-,
-
-createSection(parent, label, builder) {
-  const section = parent.createEl('div', { cls: 'apple-setting-section' });
-  section.createEl('label', { cls: 'apple-setting-label', text: label });
-  const content = section.createEl('div', { cls: 'apple-setting-content' });
-  builder(content);
-  return section;
-}
-,
-
-getEffectiveLineHeight() {
-  const configured = this.plugin.settings.lineHeight;
-  if (configured !== null && configured !== undefined) return configured;
-  const cfg = this.getThemeConfigSafe();
-  return cfg ? cfg.lineHeight : 1.8;
-}
-,
-
-getEffectiveParagraphGap() {
-  const configured = this.plugin.settings.paragraphGap;
-  if (configured !== null && configured !== undefined) return configured;
-  const cfg = this.getThemeConfigSafe();
-  return cfg ? cfg.paragraphGap : 18;
-}
-,
-
-getEffectiveLetterSpacing() {
-  const configured = this.plugin.settings.letterSpacing;
-  if (configured !== null && configured !== undefined) return configured;
-  return 0;
-}
-,
-
-getThemeConfigSafe() {
-  const theme = this.theme;
-  if (theme && typeof theme.getThemeConfig === 'function') {
-    try {
-      return theme.getThemeConfig();
-    } catch {
-      return null;
-    }
-  }
-  return null;
-}
-,
-
-formatSpacingValue(val) {
-  const num = Number(val);
-  if (!Number.isFinite(num)) return '0';
-  // 最多 2 位小数，去掉末尾多余的 0
-  return parseFloat(num.toFixed(2)).toString();
-}
-,
-
-updateSpacingSummary() {
-  const el = this.settingsSpacingValues;
-  if (!el) return;
-  const lh = this.formatSpacingValue(this.getEffectiveLineHeight());
-  const pg = this.formatSpacingValue(this.getEffectiveParagraphGap());
-  const ls = this.formatSpacingValue(this.getEffectiveLetterSpacing());
-  el.setText(`行距 ${lh} · 段距 ${pg} · 字距 ${ls}`);
-}
-,
-
-// 切换主题后刷新间距滑块显示：跟随主题(null)的滑块同步新主题默认值；
-// 用户手动设过(非 null)的保持不变（getEffective 直接返回用户值）。
-refreshSpacingSliders() {
-  if (!this.spacingSliderRefs) return;
-  this.spacingSliderRefs.forEach(({ slider, valueLabel, getEffective }) => {
-    const v = getEffective();
-    slider.value = String(v);
-    valueLabel.setText(this.formatSpacingValue(v));
-  });
-  this.updateSpacingSummary();
-}
-,
-
-resetSettingsPanelViewState() {
-  const advancedOptions = this.settingsAdvancedOptions || this.settingsOverlay?.querySelector('.apple-settings-details');
-  if (advancedOptions) advancedOptions.open = false;
-  if (this.settingsSpacingGroup) this.settingsSpacingGroup.open = false;
-
-  const scrollTargets = [
-    this.settingsOverlay,
-    this.settingsArea,
-    this.settingsAdvancedArea,
-  ].filter(Boolean);
-
-  const resetScroll = () => {
-    scrollTargets.forEach((target) => {
-      target.scrollTop = 0;
-    });
-  };
-
-  resetScroll();
-  if (typeof requestAnimationFrame === 'function') {
-    window.requestAnimationFrame(resetScroll);
-  }
-}
-,
-
-resetAiLayoutPanelViewState() {
-  this.aiAdvancedOpen = false;
-  this.aiLayoutDebugMode = '';
-  this.aiLayoutPendingAnchor = null;
-
-  const scrollTargets = [
-    this.aiLayoutOverlay,
-    this.aiLayoutArea,
-    this.aiAdvancedBody,
-    this.aiDebugPanelBody,
-  ].filter(Boolean);
-
-  const resetScroll = () => {
-    scrollTargets.forEach((target) => {
-      target.scrollTop = 0;
-    });
-  };
-
-  resetScroll();
-  if (typeof requestAnimationFrame === 'function') {
-    window.requestAnimationFrame(resetScroll);
-  }
-}
-,
-
-togglePanel(overlay, button, onOpen) {
-  if (!overlay || !button) return;
-  const willOpen = !overlay.classList.contains('visible');
-  this.closeTransientPanels();
-  if (willOpen) {
-    overlay.classList.add('visible');
-    button.classList.add('active');
-    if (typeof onOpen === 'function') onOpen();
-  }
-}
-,
-
-canScrollElementInDirection(element, deltaY) {
-  if (!element) return false;
-  const maxScroll = Math.max(0, (element.scrollHeight || 0) - (element.clientHeight || 0));
-  if (maxScroll <= 0) return false;
-  if (deltaY < 0) return (element.scrollTop || 0) > 0;
-  if (deltaY > 0) return (element.scrollTop || 0) < maxScroll - 1;
-  return true;
-}
-,
-
-attachOverlayScrollGuard(overlay, nestedSelectors = []) {
-  if (!overlay || overlay.__appleScrollGuardAttached) return;
-  const normalizedSelectors = Array.isArray(nestedSelectors)
-    ? nestedSelectors.filter(Boolean)
-    : [];
-
-  /** @param {WheelEvent} event */
-  const handleWheel = (event) => {
-    if (!overlay.classList.contains('visible')) return;
-    const target = event.target instanceof HTMLElement ? event.target : null;
-    const nestedScrollable = /** @type {Element | null} */ (target
-      ? normalizedSelectors
-        .map((selector) => target.closest(selector))
-        .find(Boolean)
-      : null);
-    const activeScrollable = nestedScrollable || overlay;
-
-    if (!this.canScrollElementInDirection(activeScrollable, event.deltaY)) {
-      event.preventDefault();
-    }
-    event.stopPropagation();
-  };
-
-  /** @param {TouchEvent} event */
-  const handleTouchMove = (event) => {
-    if (!overlay.classList.contains('visible')) return;
-    event.stopPropagation();
-  };
-
-  overlay.addEventListener('wheel', handleWheel, { passive: false });
-  overlay.addEventListener('touchmove', handleTouchMove, { passive: false });
-  overlay.__appleScrollGuardAttached = true;
-}
-,
-
-closeTransientPanels() {
-  removeElementClass(this.settingsOverlay, 'visible');
-  removeElementClass(this.aiLayoutOverlay, 'visible');
-  removeElementClass(this.settingsBtn, 'active');
-  removeElementClass(this.aiLayoutBtn, 'active');
-}
-,
-
-
-toggleSettingsPanel() {
-  const artWrapper = /** @type {unknown} */ (this.articleSettingsWrapper);
-  const stkWrapper = /** @type {unknown} */ (this.stickerSettingsWrapper);
-
-  if (this.previewMode === 'sticker') {
-    if (artWrapper && typeof artWrapper === 'object' && 'classList' in artWrapper) {
-      /** @type {Element} */ (artWrapper).classList.add('hidden');
-    }
-    if (stkWrapper && typeof stkWrapper === 'object' && 'classList' in stkWrapper) {
-      /** @type {Element} */ (stkWrapper).classList.remove('hidden');
-    }
-    const toggleState = toRecord(this.stickerIndexToggleState);
-    const checkbox = toggleState ? toRecord(toggleState.checkbox) : null;
-    if (checkbox && typeof checkbox.checked === 'boolean') {
-      checkbox.checked = Boolean(this.insertStickerImageIndex);
-    }
-  } else {
-    if (stkWrapper && typeof stkWrapper === 'object' && 'classList' in stkWrapper) {
-      /** @type {Element} */ (stkWrapper).classList.add('hidden');
-    }
-    if (artWrapper && typeof artWrapper === 'object' && 'classList' in artWrapper) {
-      /** @type {Element} */ (artWrapper).classList.remove('hidden');
-    }
-  }
-
-  this.togglePanel(this.settingsOverlay, this.settingsBtn, () => this.resetSettingsPanelViewState());
-}
-,
-
-switchPreviewMode(mode) {
-  if (this.previewMode === mode) return;
-  this.previewMode = mode;
-
-  const articleBtn = /** @type {unknown} */ (this.btnArticleMode);
-  const stickerBtn = /** @type {unknown} */ (this.btnStickerMode);
-
-  if (articleBtn && stickerBtn) {
-    const aEl = /** @type {Element} */ (articleBtn);
-    const sEl = /** @type {Element} */ (stickerBtn);
-    if (mode === 'article') {
-      aEl.classList.add('active');
-      sEl.classList.remove('active');
-    } else {
-      sEl.classList.add('active');
-      aEl.classList.remove('active');
-    }
-  }
-
-  // 跨模式切换时收起悬浮面板：文章设置与贴图设置内容不同，留在屏幕上会造成误解。
-  this.closeTransientPanels();
-
-  if (mode === 'sticker') {
-    if (this.aiLayoutBtn && typeof this.aiLayoutBtn.classList === 'object') this.aiLayoutBtn.classList.add('hidden');
-    if (this.copyBtn && typeof this.copyBtn.classList === 'object') this.copyBtn.classList.add('hidden');
-    this.renderStickerPreview();
-  } else {
-    if (this.aiLayoutBtn && typeof this.aiLayoutBtn.classList === 'object') this.aiLayoutBtn.classList.remove('hidden');
-    if (this.copyBtn && typeof this.copyBtn.classList === 'object') this.copyBtn.classList.remove('hidden');
-    this.convertCurrent(true);
-  }
-
-  const headerEl = this.containerEl ? this.containerEl.querySelector('.apple-preview-header') : null;
-  if (headerEl && this.containerEl) {
-    const h = /** @type {HTMLElement} */ (headerEl).offsetHeight || 80;
-    this.containerEl.style.setProperty('--apple-header-height', h + 'px');
-  }
-}
-,
-
-getStickerUiState(filePath) {
-  const selfRecord = toRecord(this);
-  if (!selfRecord.stickerUiStates) {
-    selfRecord.stickerUiStates = new Map();
-  }
-  const states = /** @type {Map<string, {
-   * order: string[],
-   * removedKeys: string[],
-   * manualItems: object[],
-   * undoItems: Array<{item:object,index:number,wasManual:boolean}>,
-   * objectUrls: Set<string>
-   * }>} */ (selfRecord.stickerUiStates);
-  const key = filePath || '';
-  let state = states.get(key);
-  if (!state) {
-    state = {
-      order: [],
-      removedKeys: [],
-      manualItems: [],
-      undoItems: [],
-      objectUrls: new Set(),
-    };
-    states.set(key, state);
-  }
-  return state;
-}
-,
-
-removeStickerImageItem(filePath, item, index) {
-  const uiState = this.getStickerUiState(filePath);
-  const wasManual = item?.source !== 'body';
-  uiState.undoItems.push({ item, index, wasManual });
-  if (uiState.undoItems.length > 9) uiState.undoItems.shift();
-
-  if (wasManual) {
-    uiState.manualItems = uiState.manualItems.filter((candidate) => candidate.key !== item.key);
-  } else if (!uiState.removedKeys.includes(item.key)) {
-    uiState.removedKeys.push(item.key);
-  }
-  uiState.order = uiState.order.filter((key) => key !== item.key);
-}
-,
-
-restoreLastStickerImage(filePath) {
-  const uiState = this.getStickerUiState(filePath);
-  const undo = uiState.undoItems.pop();
-  if (!undo?.item) return '';
-  const item = undo.item;
-  if (undo.wasManual && !uiState.manualItems.some((candidate) => candidate.key === item.key)) {
-    uiState.manualItems.push(item);
-  }
-  uiState.removedKeys = uiState.removedKeys.filter((key) => key !== item.key);
-  const nextOrder = uiState.order.filter((key) => key !== item.key);
-  nextOrder.splice(Math.min(Math.max(undo.index, 0), nextOrder.length), 0, item.key);
-  uiState.order = nextOrder;
-  return item.key;
-}
-,
-
-restoreAllStickerImages(filePath) {
-  const uiState = this.getStickerUiState(filePath);
-  while (uiState.undoItems.length > 0) {
-    this.restoreLastStickerImage(filePath);
-  }
-  uiState.removedKeys = [];
-}
-,
-
-resolveStickerImageSrc(src, sourcePath) {
-  const raw = String(src || '');
-  if (!raw || /^(data:|https?:\/\/|app:\/\/|capacitor:\/\/)/i.test(raw)) return raw;
-
-  const linkFile = this.app?.metadataCache?.getFirstLinkpathDest?.(raw, sourcePath || '');
-  if (!linkFile) return raw;
-
-  const vault = toRecord(this.app.vault);
-  const getResourcePathRaw = ('getResourcePath' in vault) ? vault.getResourcePath : null;
-  if (typeof getResourcePathRaw !== 'function') return raw;
-
-  const getResourcePath = /** @type {(file: unknown) => unknown} */ (getResourcePathRaw);
-  const resolved = getResourcePath(linkFile);
-  return typeof resolved === 'string' && resolved ? resolved : raw;
-}
-,
-
-async buildStickerData(options = {}) {
-  const requestedSourcePath = typeof options.sourcePath === 'string' ? options.sourcePath : '';
-  const activeFile = requestedSourcePath
-    ? (this.app?.vault?.getAbstractFileByPath?.(requestedSourcePath) || this.getPublishContextFile())
-    : this.getPublishContextFile();
-  const sourcePath = activeFile && typeof activeFile.path === 'string' ? activeFile.path : '';
-
-  // 贴图模式不复用 convertCurrent 的文章渲染结果：直接读当前笔记，
-  // 这样在贴图模式下编辑正文，预览也能实时更新。
-  let markdown = '';
-  if (requestedSourcePath && activeFile && activeFile.path === requestedSourcePath) {
-    const vault = toRecord(this.app.vault);
-    const cachedReadRaw = 'cachedRead' in vault ? vault.cachedRead : null;
-    if (typeof cachedReadRaw === 'function') {
-      markdown = String(await cachedReadRaw.call(this.app.vault, activeFile));
-    }
-  }
-  if (!markdown) {
-    const source = /** @type {MarkdownSourceResultLike} */ (await resolveMarkdownSource({
-      app: this.app,
-      lastActiveFile: this.lastActiveFile,
-      MarkdownViewType: MarkdownView,
-    }));
-    if (source.ok && typeof source.markdown === 'string') {
-      markdown = source.markdown;
-    } else if (typeof this.lastResolvedMarkdown === 'string') {
-      markdown = this.lastResolvedMarkdown;
-    }
-  }
-
-  const fileCache = activeFile ? this.app.metadataCache.getFileCache(activeFile) : null;
-  const frontmatter = (fileCache && fileCache.frontmatter && typeof fileCache.frontmatter === 'object')
-    ? /** @type {Record<string, unknown>} */ (fileCache.frontmatter)
-    : {};
-  const fallbackTitle = activeFile && typeof activeFile.basename === 'string' ? activeFile.basename : '未命名贴图';
-  const uiState = this.getStickerUiState(sourcePath);
-  const resolveBodyImageIdentity = (src) => {
-    const resolved = this.app?.metadataCache?.getFirstLinkpathDest?.(src, sourcePath);
-    return resolved && typeof resolved.path === 'string'
-      ? normalizeVaultPath(resolved.path)
-      : normalizeVaultPath(src);
-  };
-
-  const extracted = extractStickerData({
-    markdown,
-    frontmatter,
-    fallbackTitle,
-    insertImageIndex: Boolean(this.insertStickerImageIndex),
-    imageOrder: uiState.order,
-    removedImageKeys: uiState.removedKeys,
-    manualImageItems: uiState.manualItems,
-    resolveBodyImageIdentity,
-  });
-
-  // 用对齐后的结果回写用户顺序，避免下一次渲染继续携带已失效的路径。
-  uiState.order = extracted.imageItems.map((item) => item.key);
-
-  /** @type {StickerPreviewDataLike} */
-  const stickerData = {
-    title: extracted.title,
-    content: extracted.content,
-    images: extracted.images,
-    imageItems: extracted.imageItems,
-    omittedImageCount: extracted.omittedImageCount,
-    hasCodeBlocks: extracted.hasCodeBlocks,
-    hasTables: extracted.hasTables,
-    hasMath: extracted.hasMath,
-    hasFootnotes: extracted.hasFootnotes,
-    removed: extracted.removed,
-    // 弹窗与侧边栏都需要可直接显示的地址（vault 内图片必须转成 app:// 资源路径）
-    imageDisplaySources: extracted.imageItems.map((item) => (
-      item.source === 'body'
-        ? this.resolveStickerImageSrc(getStickerImageItemSrc(item), sourcePath)
-        : (item.displaySrc || getStickerImageItemSrc(item))
-    )),
-    sourcePath,
-  };
-
-  this.previewStickerData = stickerData;
-  return stickerData;
-}
-,
-
-async renderStickerPreview() {
-  if (!this.previewContainer) return undefined;
-
-  // 连续输入会触发多次渲染：只让最后一次结果落到 DOM，避免卡片重复堆叠。
-  const selfRecord = toRecord(this);
-  const generation = (typeof selfRecord.stickerRenderGeneration === 'number' ? selfRecord.stickerRenderGeneration : 0) + 1;
-  selfRecord.stickerRenderGeneration = generation;
-
-  const stickerData = await this.buildStickerData();
-  if (this.previewMode !== 'sticker' || selfRecord.stickerRenderGeneration !== generation) return undefined;
-
-  this.previewContainer.empty();
-  const stickerContainer = this.previewContainer.createEl('div', { cls: 'apple-sticker-preview-wrapper' });
-  const uiState = this.getStickerUiState(stickerData.sourcePath || '');
-
-  // 1. 被过滤内容提醒（代码块 / 表格在微信贴图里无法呈现）
-  const strippedParts = (Array.isArray(stickerData.removed) ? stickerData.removed : [])
-    .filter((entry) => entry?.count > 0)
-    .map((entry) => `${entry.kind} ${entry.count} 处`);
-  if (strippedParts.length > 0) {
-    const notice = stickerContainer.createEl('div', { cls: 'apple-sticker-notice-warning' });
-    const noticeIcon = notice.createEl('span', { cls: 'apple-sticker-notice-icon' });
-    const noticeSetIcon = getObsidianSetIcon();
-    if (typeof noticeSetIcon === 'function') noticeSetIcon(noticeIcon, 'info');
-    const noticeContent = notice.createEl('div', { cls: 'apple-sticker-notice-content' });
-    noticeContent.createEl('span', { cls: 'apple-sticker-notice-title', text: `已清理：${strippedParts.join('、')}` });
-    noticeContent.createEl('span', {
-      cls: 'apple-sticker-notice-desc',
-      text: '清理只影响贴图草稿，不会改动笔记原文。',
-    });
-  }
-
-  // 2. 图片卡片墙预览与交互（上限 20 张）
-  const imagesSection = stickerContainer.createEl('div', { cls: 'apple-sticker-images-section' });
-  const sectionHeader = imagesSection.createEl('div', { cls: 'apple-sticker-section-header' });
-  const sectionTitle = sectionHeader.createEl('div', { cls: 'apple-sticker-section-title' });
-  const setIcon = getObsidianSetIcon();
-  if (typeof setIcon === 'function') {
-    const iconSpan = sectionTitle.createEl('span', { cls: 'apple-sticker-section-icon-lucide' });
-    setIcon(iconSpan, 'image');
-  }
-  sectionTitle.createEl('span', { text: '贴图图片列表' });
-  sectionHeader.createEl('span', {
-    cls: 'apple-sticker-count-badge',
-    text: `${stickerData.imageItems.length} / ${STICKER_MAX_IMAGES} 张`
-  });
-
-  renderStickerImageList(imagesSection, {
-    items: stickerData.imageItems,
-    getDisplaySrc: (_, index) => stickerData.imageDisplaySources[index] || '',
-    setIcon,
-    emptyText: '正文中还没有图片；可在发布弹窗中添加本地图片或公众号素材。',
-    onMove: (fromIndex, toIndex, movedItem) => {
-      uiState.order = moveStickerImageItem(
-        stickerData.imageItems.map((item) => item.key),
-        fromIndex,
-        toIndex
-      );
-      selfRecord.stickerFocusKey = movedItem?.key || '';
-      void this.renderStickerPreview();
-    },
-    onRemove: (item, index) => {
-      this.removeStickerImageItem(stickerData.sourcePath, item, index);
-      void this.renderStickerPreview();
-    },
-    focusKey: typeof selfRecord.stickerFocusKey === 'string' ? selfRecord.stickerFocusKey : '',
-  });
-  selfRecord.stickerFocusKey = '';
-
-  if (uiState.undoItems.length > 0 || uiState.removedKeys.length > 0) {
-    const restoreRow = imagesSection.createDiv({ cls: 'apple-sticker-restore-row' });
-    if (uiState.undoItems.length > 0) {
-      const undoButton = restoreRow.createEl('button', {
-        cls: 'apple-sticker-restore-btn',
-        text: '撤销上次移除',
-      });
-      undoButton.onclick = () => {
-        selfRecord.stickerFocusKey = this.restoreLastStickerImage(stickerData.sourcePath);
-        void this.renderStickerPreview();
-      };
-    }
-    const restoreAllButton = restoreRow.createEl('button', {
-      cls: 'apple-sticker-restore-btn',
-      text: '恢复全部',
-    });
-    restoreAllButton.onclick = () => {
-      this.restoreAllStickerImages(stickerData.sourcePath);
-      void this.renderStickerPreview();
-    };
-  }
-
-  // 💡 温馨提示：贴图提取自正文，如需添加请在发布弹窗操作
-  const hintLine = imagesSection.createEl('div', { cls: 'apple-sticker-hint-line' });
-  const hintIcon = hintLine.createSpan({ cls: 'apple-sticker-hint-icon' });
-  if (typeof setIcon === 'function') {
-    setIcon(hintIcon, 'info');
-  }
-  hintLine.createEl('span', {
-    cls: 'apple-sticker-hint-text',
-    text: '这里确认发布顺序；添加本地图片或公众号素材，请打开发布弹窗。移除不会改动笔记。'
-  });
-
-  // 3. 纯文本正文预览
-  const textSection = stickerContainer.createEl('div', { cls: 'apple-sticker-text-section' });
-  const textHeader = textSection.createEl('div', { cls: 'apple-sticker-section-header' });
-  const textTitle = textHeader.createEl('div', { cls: 'apple-sticker-section-title' });
-  if (typeof setIcon === 'function') {
-    const iconSpan = textTitle.createEl('span', { cls: 'apple-sticker-section-icon-lucide' });
-    setIcon(iconSpan, 'file-text');
-  }
-  textTitle.createEl('span', { text: '贴图描述文案' });
-  
-  const charCount = stickerData.content ? stickerData.content.length : 0;
-  const isOverLimit = charCount > STICKER_MAX_CONTENT_LENGTH;
-  const countBadge = textHeader.createEl('span', { cls: 'apple-sticker-count-badge' });
-  countBadge.createEl('span', {
-    cls: `apple-sticker-count-current${isOverLimit ? ' is-error' : ''}`,
-    text: `${charCount}`
-  });
-  countBadge.createEl('span', {
-    cls: 'apple-sticker-count-total',
-    text: ` / ${STICKER_MAX_CONTENT_LENGTH}`
-  });
-
-  textSection.createEl('div', {
-    cls: 'apple-sticker-text-preview',
-    text: stickerData.content || '(正文无纯文本内容)'
-  });
-
-  return stickerContainer;
-}
-,
-};
